0.5.0
The upgrade to fully integrate the new scheme required a bunch of changes that more or less broke everything.

+++ LaserOperation.py
@@ -25,7 +25,7 @@ class LaserOperation(list):
 
         power = 1000
         if 'power' in obj.values and obj.values['power'] is not None:
-            power = int(obj.values['power'])
+            power = float(obj.values['power'])
         self.power = power
 
         dratio = None
